refactor folders and add getter for questions in quiz
