+++ src/interface_adapter/quiz/take_quiz/TakeQuizController.java
@@ -1,0 +1,2 @@
+package interface_adapter.quiz.take_quiz;public class TakeQuizController {
+}
